refactor: changed js to tsx
